+++ java/com/grammatech/gtirb/SymAddrAddr.java
@@ -144,12 +144,13 @@ public class SymAddrAddr extends SymbolicExpression {
             Util.uuidToByteString(this.getSymbol2Uuid()));
         protoSymAddrAddr.setOffset(this.getOffset());
         protoSymbolicExpression.setAddrAddr(protoSymAddrAddr);
-
-        // NOTE for this to be valid, a one-to-one mapping of enums must be
-        // maintained
-        for (AttributeFlag attributeFlag : this.getAttributeFlags())
+        for (AttributeFlag attributeFlag : this.getAttributeFlags()) {
             protoSymbolicExpression.addAttributeFlagsValue(
-                attributeFlag.ordinal());
+                attributeFlag.value());
+        }
+        for (Integer value : this.getUnknownAttributeFlags()) {
+            protoSymbolicExpression.addAttributeFlagsValue(value);
+        }
         return protoSymbolicExpression;
     }
 }