+++ java/com/grammatech/gtirb/SymbolicExpression.java
@@ -16,7 +16,10 @@ package com.grammatech.gtirb;
 
 import com.grammatech.gtirb.proto.SymbolicExpressionOuterClass;
 import java.util.ArrayList;
+import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 /**
  * The Symbolic Expression class is a base class for expressions such as
@@ -25,38 +28,101 @@ import java.util.List;
 public class SymbolicExpression implements TreeListItem {
 
     /**
-     * 	Attributes for Symbolic Expressions
+     * Attributes for Symbolic Expressions.
+     *
+     * @see https://grammatech.github.io/gtirb/md__symbolic_expression.html
      */
     public enum AttributeFlag {
-        Part0,
-        Part1,
-        Part2,
-        Part3,
-        Adjusted,
-        GotRef,
-        GotRelPC,
-        GotRelGot,
-        AddrRelGot,
-        GotRelAddr,
-        GotPage,
-        GotPageOfst,
-        PltRef,
-        GotOff,
-        TpOff,
-        NtpOff,
-        DtpOff,
-        TlsGd,
-        TlsLd,
-        Lo,
-        Hi,
-        Lo12,
-        Abs,
-        Signed,
-        NoOverflowCheck
+        GOT(SymbolicExpressionOuterClass.SymAttribute.GOT_VALUE),
+        GOTPC(SymbolicExpressionOuterClass.SymAttribute.GOTPC_VALUE),
+        GOTOFF(SymbolicExpressionOuterClass.SymAttribute.GOTOFF_VALUE),
+        GOTREL(SymbolicExpressionOuterClass.SymAttribute.GOTREL_VALUE),
+        PLT(SymbolicExpressionOuterClass.SymAttribute.PLT_VALUE),
+        PLTOFF(SymbolicExpressionOuterClass.SymAttribute.PLTOFF_VALUE),
+        PCREL(SymbolicExpressionOuterClass.SymAttribute.PCREL_VALUE),
+        SECREL(SymbolicExpressionOuterClass.SymAttribute.SECREL_VALUE),
+        TLS(SymbolicExpressionOuterClass.SymAttribute.TLS_VALUE),
+        TLSGD(SymbolicExpressionOuterClass.SymAttribute.TLSGD_VALUE),
+        TLSLD(SymbolicExpressionOuterClass.SymAttribute.TLSLD_VALUE),
+        TLSLDM(SymbolicExpressionOuterClass.SymAttribute.TLSLDM_VALUE),
+        TLSCALL(SymbolicExpressionOuterClass.SymAttribute.TLSCALL_VALUE),
+        TLSDESC(SymbolicExpressionOuterClass.SymAttribute.TLSDESC_VALUE),
+        TPREL(SymbolicExpressionOuterClass.SymAttribute.TPREL_VALUE),
+        TPOFF(SymbolicExpressionOuterClass.SymAttribute.TPOFF_VALUE),
+        DTPREL(SymbolicExpressionOuterClass.SymAttribute.DTPREL_VALUE),
+        DTPOFF(SymbolicExpressionOuterClass.SymAttribute.DTPOFF_VALUE),
+        DTPMOD(SymbolicExpressionOuterClass.SymAttribute.DTPMOD_VALUE),
+        NTPOFF(SymbolicExpressionOuterClass.SymAttribute.NTPOFF_VALUE),
+        PAGE(SymbolicExpressionOuterClass.SymAttribute.PAGE_VALUE),
+        PAGEOFF(SymbolicExpressionOuterClass.SymAttribute.PAGEOFF_VALUE),
+        CALL(SymbolicExpressionOuterClass.SymAttribute.CALL_VALUE),
+        LO(SymbolicExpressionOuterClass.SymAttribute.LO_VALUE),
+        HI(SymbolicExpressionOuterClass.SymAttribute.HI_VALUE),
+        HIGHER(SymbolicExpressionOuterClass.SymAttribute.HIGHER_VALUE),
+        HIGHEST(SymbolicExpressionOuterClass.SymAttribute.HIGHEST_VALUE),
+        GOTNTPOFF(SymbolicExpressionOuterClass.SymAttribute.GOTNTPOFF_VALUE),
+        INDNTPOFF(SymbolicExpressionOuterClass.SymAttribute.INDNTPOFF_VALUE),
+        G0(SymbolicExpressionOuterClass.SymAttribute.G0_VALUE),
+        G1(SymbolicExpressionOuterClass.SymAttribute.G1_VALUE),
+        G2(SymbolicExpressionOuterClass.SymAttribute.G2_VALUE),
+        G3(SymbolicExpressionOuterClass.SymAttribute.G3_VALUE),
+        UPPER16(SymbolicExpressionOuterClass.SymAttribute.UPPER16_VALUE),
+        LOWER16(SymbolicExpressionOuterClass.SymAttribute.LOWER16_VALUE),
+        LO12(SymbolicExpressionOuterClass.SymAttribute.LO12_VALUE),
+        LO15(SymbolicExpressionOuterClass.SymAttribute.LO15_VALUE),
+        LO14(SymbolicExpressionOuterClass.SymAttribute.LO14_VALUE),
+        HI12(SymbolicExpressionOuterClass.SymAttribute.HI12_VALUE),
+        HI21(SymbolicExpressionOuterClass.SymAttribute.HI21_VALUE),
+        S(SymbolicExpressionOuterClass.SymAttribute.S_VALUE),
+        PG(SymbolicExpressionOuterClass.SymAttribute.PG_VALUE),
+        NC(SymbolicExpressionOuterClass.SymAttribute.NC_VALUE),
+        ABS(SymbolicExpressionOuterClass.SymAttribute.ABS_VALUE),
+        PREL(SymbolicExpressionOuterClass.SymAttribute.PREL_VALUE),
+        PREL31(SymbolicExpressionOuterClass.SymAttribute.PREL31_VALUE),
+        TARGET1(SymbolicExpressionOuterClass.SymAttribute.TARGET1_VALUE),
+        TARGET2(SymbolicExpressionOuterClass.SymAttribute.TARGET2_VALUE),
+        SBREL(SymbolicExpressionOuterClass.SymAttribute.SBREL_VALUE),
+        TLSLDO(SymbolicExpressionOuterClass.SymAttribute.TLSLDO_VALUE),
+        HI16(SymbolicExpressionOuterClass.SymAttribute.HI16_VALUE),
+        LO16(SymbolicExpressionOuterClass.SymAttribute.LO16_VALUE),
+        GPREL(SymbolicExpressionOuterClass.SymAttribute.GPREL_VALUE),
+        DISP(SymbolicExpressionOuterClass.SymAttribute.DISP_VALUE),
+        OFST(SymbolicExpressionOuterClass.SymAttribute.OFST_VALUE),
+        H(SymbolicExpressionOuterClass.SymAttribute.H_VALUE),
+        L(SymbolicExpressionOuterClass.SymAttribute.L_VALUE),
+        HA(SymbolicExpressionOuterClass.SymAttribute.HA_VALUE),
+        HIGH(SymbolicExpressionOuterClass.SymAttribute.HIGH_VALUE),
+        HIGHA(SymbolicExpressionOuterClass.SymAttribute.HIGHA_VALUE),
+        HIGHERA(SymbolicExpressionOuterClass.SymAttribute.HIGHERA_VALUE),
+        HIGHESTA(SymbolicExpressionOuterClass.SymAttribute.HIGHESTA_VALUE),
+        TOCBASE(SymbolicExpressionOuterClass.SymAttribute.TOCBASE_VALUE),
+        TOC(SymbolicExpressionOuterClass.SymAttribute.TOC_VALUE),
+        NOTOC(SymbolicExpressionOuterClass.SymAttribute.NOTOC_VALUE),
+        ;
+
+        private final int value;
+        private static final Map<Integer, AttributeFlag> mapping = initMap();
+
+        AttributeFlag(int value) { this.value = value; }
+
+        public int value() { return this.value; }
+
+        public static AttributeFlag fromInteger(int value) {
+            return mapping.get(value);
+        }
+
+        private static Map<Integer, AttributeFlag> initMap() {
+            Map<Integer, AttributeFlag> mapping = new HashMap<>();
+            for (AttributeFlag flag : AttributeFlag.values()) {
+                mapping.put(flag.value(), flag);
+            }
+            return Collections.unmodifiableMap(mapping);
+        }
     }
 
     private long offset;
     private List<AttributeFlag> attributeFlags;
+    private List<Integer> unknownAttributeFlags;
 
     // This is the constructor used when instantiating a sub class
     // NOTE: Offset is not set in constructor because subclasss doesn't
@@ -70,10 +136,15 @@ public class SymbolicExpression implements TreeListItem {
     public SymbolicExpression(SymbolicExpressionOuterClass
                                   .SymbolicExpression protoSymbolicExpression) {
         this.attributeFlags = new ArrayList<AttributeFlag>();
+        this.unknownAttributeFlags = new ArrayList<Integer>();
         for (Integer value :
              protoSymbolicExpression.getAttributeFlagsValueList()) {
-            AttributeFlag attributeFlag = AttributeFlag.values()[value];
-            this.attributeFlags.add(attributeFlag);
+            AttributeFlag attributeFlag = AttributeFlag.fromInteger(value);
+            if (attributeFlag == null) {
+                this.unknownAttributeFlags.add(value);
+            } else {
+                this.attributeFlags.add(attributeFlag);
+            }
         }
     }
 
@@ -85,6 +156,7 @@ public class SymbolicExpression implements TreeListItem {
     public SymbolicExpression(long offset, List<AttributeFlag> attributeFlags) {
         this.setOffset(offset);
         this.setAttributeFlags(attributeFlags);
+        this.setUnknownAttributeFlags(new ArrayList<Integer>());
     }
 
     /**
@@ -140,6 +212,25 @@ public class SymbolicExpression implements TreeListItem {
      */
     public void setAttributeFlags(List<AttributeFlag> attributeFlags) {
         this.attributeFlags = attributeFlags;
+    }
+
+    /**
+     * Get unknown attribute flags applying to this SymbolicExpression.
+     *
+     * @return  A set of flags applying to this symbolic expression.
+     */
+    public List<Integer> getUnknownAttributeFlags() {
+        return this.unknownAttributeFlags;
+    }
+
+    /**
+     * Set the attribute flags of this SymbolicExpression.
+     *
+     * @param attributeFlags    A set of flags that will be applied to this
+     * symbolic expression.
+     */
+    public void setUnknownAttributeFlags(List<Integer> attributeFlags) {
+        this.unknownAttributeFlags = attributeFlags;
     }
 
     /**