+++ python/tests/test_deep_eq.py
@@ -109,12 +109,12 @@ class DeepEqTest(unittest.TestCase):
         s1 = gtirb.SymAddrConst(
             offset=1,
             symbol=gtirb.Symbol(name="name", payload=None, uuid=id1),
-            attributes={gtirb.SymbolicExpression.Attribute.Part1},
+            attributes={gtirb.SymbolicExpression.Attribute.G1},
         )
         s2 = gtirb.SymAddrConst(
             offset=1,
             symbol=gtirb.Symbol(name="name", payload=None, uuid=id1),
-            attributes={gtirb.SymbolicExpression.Attribute.Part1},
+            attributes={gtirb.SymbolicExpression.Attribute.G1},
         )
         self.assertTrue(s1.deep_eq(s2))
 
@@ -137,7 +137,7 @@ class DeepEqTest(unittest.TestCase):
         s1 = gtirb.SymAddrConst(
             offset=1,
             symbol=gtirb.Symbol(name="name", payload=None, uuid=id1),
-            attributes={gtirb.SymbolicExpression.Attribute.Part1},
+            attributes={gtirb.SymbolicExpression.Attribute.G1},
         )
         s2 = gtirb.SymAddrConst(
             offset=1,
@@ -151,14 +151,14 @@ class DeepEqTest(unittest.TestCase):
             scale=2,
             symbol1=gtirb.Symbol(name="name1", payload=None, uuid=id1),
             symbol2=gtirb.Symbol(name="name2", payload=None, uuid=id2),
-            attributes={gtirb.SymbolicExpression.Attribute.Part1},
+            attributes={gtirb.SymbolicExpression.Attribute.G1},
         )
         s2 = gtirb.SymAddrAddr(
             offset=1,
             scale=2,
             symbol1=gtirb.Symbol(name="name1", payload=None, uuid=id1),
             symbol2=gtirb.Symbol(name="name2", payload=None, uuid=id2),
-            attributes={gtirb.SymbolicExpression.Attribute.Part1},
+            attributes={gtirb.SymbolicExpression.Attribute.G1},
         )
         self.assertTrue(s1.deep_eq(s2))
 
@@ -223,7 +223,7 @@ class DeepEqTest(unittest.TestCase):
             scale=2,
             symbol1=gtirb.Symbol(name="name1", payload=None, uuid=id1),
             symbol2=gtirb.Symbol(name="name2", payload=None, uuid=id2),
-            attributes={gtirb.SymbolicExpression.Attribute.Part1},
+            attributes={gtirb.SymbolicExpression.Attribute.G1},
         )
         s2 = gtirb.SymAddrAddr(
             offset=1,