+++ src/test/SymbolicExpression.test.cpp
@@ -31,8 +31,7 @@ TEST(Unit_SymbolicExpression, protobufRoundTrip) {
   // SymAddrConst
   {
     SymAttributeSet OrigSASet;
-    OrigSASet.addFlags(SymAttribute::Part0, SymAttribute::Adjusted,
-                       SymAttribute::PltRef);
+    OrigSASet.insert({SymAttribute::ABS, SymAttribute::G0});
     SymbolicExpression original(SymAddrConst{1, Sym1, OrigSASet});
 
     gtirb::SymbolicExpression Result;
@@ -63,76 +62,38 @@ TEST(Unit_SymbolicExpression, protobufRoundTrip) {
   }
 }
 
-TEST(Unit_SymAttributeSet, Base) {
-  gtirb::SymAttributeSet SASet;
+TEST(Unit_SymAttributeSet, unknownAttributes) {
+  std::ostringstream Out;
+  {
+    auto* I = IR::Create(Ctx);
+    auto* M = I->addModule(Ctx, "foo");
+    auto* S = M->addSection(Ctx, "bar");
+    auto* B = S->addByteInterval(Ctx, Addr(0), 8);
 
-  EXPECT_EQ(SASet.begin(), SASet.end());
-  for (size_t I = 0; I <= static_cast<size_t>(SymAttribute::Max); ++I) {
-    EXPECT_FALSE(SASet.isFlagSet(static_cast<SymAttribute>(I)));
+    Symbol* Sym = Symbol::Create(Ctx, Addr(4), "baz");
+    SymAttributeSet Attrs{SymAttribute::GOT, static_cast<SymAttribute>(0xBEEF)};
+    B->addSymbolicExpression<SymAddrConst>(4, SymAddrConst{0, Sym, Attrs});
+
+    I->save(Out);
   }
 
-  // Note: using Max here instead of PltRef so test continues to
-  // work if more attributes are added.
-  SASet.addFlags(SymAttribute::Part0, SymAttribute::Adjusted,
-                 SymAttribute::Max);
-  EXPECT_TRUE(SASet.isFlagSet(SymAttribute::Part0));
-  EXPECT_TRUE(SASet.isFlagSet(SymAttribute::Adjusted));
-  EXPECT_TRUE(SASet.isFlagSet(SymAttribute::Max));
-  EXPECT_FALSE(SASet.isFlagSet(SymAttribute::Part1));
+  std::istringstream In(Out.str());
+  {
+    auto ResultOrErr = IR::load(Ctx, In);
+    ASSERT_TRUE(ResultOrErr);
 
-  SASet.removeFlag(SymAttribute::Part0);
-  EXPECT_FALSE(SASet.isFlagSet(SymAttribute::Part0));
+    auto* I = *ResultOrErr;
+    EXPECT_EQ(std::distance(I->symbolic_expressions_begin(),
+                            I->symbolic_expressions_end()),
+              1);
+    EXPECT_TRUE(std::holds_alternative<SymAddrConst>(
+        I->symbolic_expressions_begin()->getSymbolicExpression()));
 
-  SASet.addFlag(SymAttribute::Part1);
-  EXPECT_TRUE(SASet.isFlagSet(SymAttribute::Part1));
-
-  bool SawPart1 = false;
-  bool SawAdj = false;
-  bool SawMax = false;
-  bool SawUnexpected = false;
-  for (auto It = SASet.flags().begin(); It != SASet.flags().end(); ++It) {
-    switch (*It) {
-    case SymAttribute::Part1:
-      SawPart1 = true;
-      break;
-    case SymAttribute::Adjusted:
-      SawAdj = true;
-      break;
-    case SymAttribute::Max:
-      SawMax = true;
-      break;
-    default:
-      SawUnexpected = true;
-    }
+    SymAttributeSet Attrs =
+        std::get<SymAddrConst>(
+            I->symbolic_expressions_begin()->getSymbolicExpression())
+            .Attributes;
+    EXPECT_EQ(Attrs.count(SymAttribute::GOT), 1);
+    EXPECT_EQ(Attrs.count(static_cast<SymAttribute>(0xBEEF)), 1);
   }
-  EXPECT_TRUE(SawPart1);
-  EXPECT_TRUE(SawAdj);
-  EXPECT_TRUE(SawMax);
-  EXPECT_FALSE(SawUnexpected);
-
-  SawPart1 = false;
-  SawAdj = false;
-  SawMax = false;
-  auto It = SASet.end();
-  do {
-    --It;
-    ASSERT_NE(It, SASet.flags().end());
-    switch (*It) {
-    case SymAttribute::Part1:
-      SawPart1 = true;
-      break;
-    case SymAttribute::Adjusted:
-      SawAdj = true;
-      break;
-    case SymAttribute::Max:
-      SawMax = true;
-      break;
-    default:
-      SawUnexpected = true;
-    }
-  } while (It != SASet.begin());
-  EXPECT_TRUE(SawPart1);
-  EXPECT_TRUE(SawAdj);
-  EXPECT_TRUE(SawMax);
-  EXPECT_FALSE(SawUnexpected);
 }